+++ integrations/nielsen-dcr/HISTORY.md
@@ -1,3 +1,28 @@
+2.0.0 / 2020-02-06
+==================
+
+  * Events are mapped to Nielsen DCR only when Segment `load_type` property is set to 'dynamic'.
+
+1.5.0 / 2020-01-27
+==================
+
+  * Supports tracking videos in SPA apps.
+
+1.4.0 / 2020-01-17
+==================
+
+  * Change approach to persisting `content_asset_id` to better keep track of session-level info.
+
+1.3.5 / 2020-01-06
+==================
+
+  * Send unix timestamp in seconds (rather than milliseconds) when livestreams end.
+
+1.3.4 / 2019-10-23
+==================
+
+  * Default content length to `0` if not explicitly defined.
+
 1.3.3 / 2019-10-11
 ==================